* edit posix/time.c to avoid clobbering PR #1251
* delete moved files & directories
