--- nativelib/src/main/resources/limits.c
@@ -1,3 +1,0 @@
-#include <limits.h>
-
-int scalanative_path_max() { return NAME_MAX; }
