+++ nativelib/src/main/resources/posix/dirent.c
@@ -1,12 +1,43 @@
 #include <dirent.h>
 #include <string.h>
+#include <stdio.h>
+#include <sys/errno.h>
 
 #define NAME_MAX 255
 
 struct scalanative_dirent {
     unsigned long long d_ino;  /** file serial number */
     char d_name[NAME_MAX + 1]; /** name of entry */
+    short d_type;
 };
+
+int scalanative_dt_unknown() {
+  return DT_UNKNOWN;
+}
+int scalanative_dt_fifo() {
+  return DT_FIFO;
+}
+int scalanative_dt_chr() {
+  return DT_CHR;
+}
+int scalanative_dt_dir() {
+  return DT_DIR;
+}
+int scalanative_dt_blk() {
+  return DT_BLK;
+}
+int scalanative_dt_reg() {
+  return DT_REG;
+}
+int scalanative_dt_lnk() {
+  return DT_LNK;
+}
+int scalanative_dt_sock() {
+  return DT_SOCK;
+}
+int scalanative_dt_wht() {
+  return DT_WHT;
+}
 
 DIR *scalanative_opendir(const char *name) { return opendir(name); }
 
@@ -15,6 +46,7 @@ void scalanative_dirent_init(struct dirent *dirent,
     my_dirent->d_ino = dirent->d_ino;
     strncpy(my_dirent->d_name, dirent->d_name, NAME_MAX);
     my_dirent->d_name[NAME_MAX] = '\0';
+    my_dirent->d_type = dirent->d_type;
 }
 
 int scalanative_readdir(DIR *dirp, struct scalanative_dirent *buf) {
@@ -23,7 +55,12 @@ int scalanative_readdir(DIR *dirp, struct scalanative_dirent *buf) {
         scalanative_dirent_init(orig_buf, buf);
         return 0;
     } else {
-        return 1;
+        switch(errno) {
+          case EBADF: return EBADF;
+          case EFAULT: return EFAULT;
+          case EIO: return EIO;
+        }
+        return -1;
     }
 }
