+++ nativelib/src/main/resources/posix/time.c
@@ -94,8 +94,8 @@ size_t scalanative_strftime(char *buf, size_t maxsize, const char *format,
 long long scalanative_current_time_millis() {
     long long current_time_millis;
 
-#define MILLIS_PER_SEC 1000
-#define MICROS_PER_MILLI 1000
+#define MILLIS_PER_SEC 1000LL
+#define MICROS_PER_MILLI 1000LL
 
     struct timeval tv;
     gettimeofday(&tv, NULL);
