--- nativelib/src/main/resources/utime.c
@@ -1,5 +1,0 @@
-#include <utime.h>
-
-int scalanative_utime(char *path, struct utimbuf *times) {
-    return utime(path, times);
-}
